+++ cmd/exec.go
@@ -12,8 +12,8 @@ import (
 	"github.com/Azure/acr-builder/builder"
 	"github.com/Azure/acr-builder/graph"
 	"github.com/Azure/acr-builder/pkg/taskmanager"
+	"github.com/Azure/acr-builder/pkg/volume"
 	"github.com/Azure/acr-builder/templating"
-	"github.com/Azure/acr-builder/volume"
 	"github.com/google/uuid"
 	"github.com/spf13/cobra"
 )
